# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Pokémon Shiny Encounter Counter — REST API",
        "title": "Encounty API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/backgrounds/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backgrounds"
                ],
                "summary": "Upload a background image",
                "parameters": [
                    {
                        "description": "Base64-encoded image",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.BackgroundUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.FilenameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/detector/screenshot": {
            "get": {
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "detector"
                ],
                "summary": "Capture primary monitor screenshot",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/games/sync": {
            "post": {
                "description": "Triggers a background sync of game metadata from PokeAPI",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Sync games from PokeAPI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.GamesSyncResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/hotkeys": {
            "post": {
                "description": "Replaces the full hotkey map and re-registers all bindings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotkeys"
                ],
                "summary": "Update all hotkeys",
                "parameters": [
                    {
                        "description": "Complete hotkey map",
                        "name": "hotkeys",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.HotkeyMap"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.HotkeyMap"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/pokedex": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokedex"
                ],
                "summary": "Get the Pokedex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_server.PokedexEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/pokemon": {
            "post": {
                "description": "Creates a new Pokemon entry with a generated UUID and timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokemon"
                ],
                "summary": "Create a new Pokemon",
                "parameters": [
                    {
                        "description": "Pokemon to create",
                        "name": "pokemon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.Pokemon"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.Pokemon"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/restore": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Restore from a backup ZIP",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Backup ZIP file",
                        "name": "backup",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.RestoreResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "post": {
                "description": "Replaces the settings block and reconfigures file output",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Updated settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.Settings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/settings/config-path": {
            "post": {
                "description": "Moves all data to a new directory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Set config directory path",
                "parameters": [
                    {
                        "description": "New config path",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.SetConfigPathRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.PathResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/stats/overview": {
            "get": {
                "description": "Returns global encounter statistics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get statistics overview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_database.OverviewStats"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/sync/pokemon": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokedex"
                ],
                "summary": "Sync Pokedex from PokeAPI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.PokedexSyncResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/update/apply": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Apply a downloaded update",
                "parameters": [
                    {
                        "description": "Download URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.UpdateApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/update/check": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Check for available updates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.UpdateInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_zsleyer_encounty_backend_internal_state.HotkeyMap": {
            "type": "object",
            "properties": {
                "decrement": {
                    "type": "string"
                },
                "increment": {
                    "type": "string"
                },
                "next_pokemon": {
                    "type": "string"
                },
                "reset": {
                    "type": "string"
                }
            }
        },
...
        "github_com_zsleyer_encounty_backend_internal_state.Pokemon": {
            "type": "object",
            "properties": {
                "canonical_name": {
                    "description": "English PokéAPI slug",
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "detector_config": {
                    "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.DetectorConfig"
                },
                "encounters": {
                    "type": "integer"
                },
                "game": {
                    "description": "key from games.json",
                    "type": "string"
                },
                "hunt_mode": {
                    "description": "\"both\" | \"timer\" | \"detector\" (default \"both\")",
                    "type": "string"
                },
                "hunt_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "language": {
                    "description": "\"de\" | \"en\"",
                    "type": "string"
                },
                "name": {
                    "description": "Display name (localized)",
                    "type": "string"
                },
                "overlay": {
                    "description": "Pokemon-specific overlay settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_zsleyer_encounty_backend_internal_state.OverlaySettings"
                        }
                    ]
                },
                "overlay_mode": {
                    "description": "\"default\" | \"custom\" | \"linked:\u003cpokemon-id\u003e\"",
                    "type": "string"
                },
                "sprite_style": {
                    "description": "\"classic\" | \"animated\" | \"3d\" | \"artwork\"",
                    "type": "string"
                },
                "sprite_type": {
                    "description": "\"normal\" | \"shiny\"",
                    "type": "string"
                },
                "sprite_url": {
                    "type": "string"
                },
                "step": {
                    "description": "Increment/decrement step size (default 1)",
                    "type": "integer"
                },
                "timer_accumulated_ms": {
                    "type": "integer"
                },
                "timer_started_at": {
                    "type": "string"
                },
                "title": {
                    "description": "User-defined custom title",
                    "type": "string"
                }
            }
        },
...
        "internal_server.BackgroundUploadRequest": {
            "type": "object",
            "properties": {
                "image_base64": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.FilenameResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.GamesSyncResult": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
...
        "internal_server.PathResponse": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "internal_server.PokedexEntry": {
            "type": "object",
            "properties": {
                "canonical": {
                    "type": "string"
                },
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server.PokemonForm"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "names": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_server.RestoreResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "internal_server.SetConfigPathRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.errResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}